fix rename component
